First pass at shield wall.
First attempt at implementing the shield wall structure.

I had to jig around where the collision detection is done to an awkward
place in the middle of the sector code.

Currently not fully functional as for some reason walls are not
deactivating when out of power. This should be the first thing to fix
next.