Shield Wall 1.0
Shield walls first pass completed.

I had to significantly rework the pathfinding code to return resources
even if the complete request couldn't be finished.

This also required a slight reworking of the structure manager code to
decouple the resource variable reset to be before the weapons collision
checks.

This has introduced several required reworkings;

1. All structures working on the old code will think the resource
request has failed if it isn't 100% completed despite resources being
taken even if it is 99% complete. To fix add extra checks to see if it
partially completed

2. Connection lines won't light up correctly if resources are gathered
from multiple sources. This is what I'm working on next.