Stats 1.0
All objects currently in game draw a box around themselves and display
their health if they're a structure or a ship.

All objects now also draw as a target in the top right corner.

If they are an asteroid they simply draw the box at present time. It
would be very easy to add additional information.

Currently ship components call all of the relevant code, but do not draw
a box as they don't have a radius. This is a waste of time to "fix" at
the moment as the component code has to undergo changes anyway to centre
itself around its own middle.

Adding a button or series of buttons is possible and relatively easy.