Input 1.0
After testing on mobile and tablet I've come to the unfortunate
conclusion that it is simply a step too far for this virtical slice.
Huge fundamental changes are required to get it to run even remotely
satisfactorily on those platforms.

Much later on in development of this VS it might be possible to make a
mildly playable version running at 5-10 FPS but nothing reprisentitive,
more of a token effort.

I have therefore removed the vast majority of touch controls and
reverted back to a keyboard and mouse. The code is still there and could
be reimplemented, but for now it is completely disabled.