Cargo 0.1
Very happy with progress on cargo holds.

Objects can be dragged and dropped into the cargo hold. Lots of bugs
squashed and the code isn't very readable, but it is functional and
extendable.

Checks on capacity still need to be performed and a maximum distance for
storing and ejecting cargo needs to be set. Right clicking to
instant-store is also needed for 1.0.