Production 0.2
Second pass at production.

Blueprints now exist, have a number of runs. Each gameobject has a cost
and time to be produced and the assembler automatically deducts the
correct amount of metal.

Next task is to produce "packed" version of ships when construction is
complete... then go on to allow players to spawn as those ships. Should
be acheived by tomorrow.